Modify rules of tests

+++ src/hooks/useCart.tsx
@@ -22,7 +22,8 @@ interface CartContextData {
 const CartContext = createContext<CartContextData>({} as CartContextData);
 
 export function CartProvider({ children }: CartProviderProps): JSX.Element {
-  const [stock, setStock] = useState<Stock[]>([]);
+  const [stocks, setStocks] = useState<Stock[]>([]);
+  const [products, setProducts] = useState<Product[]>([]);
   const [cart, setCart] = useState<Product[]>(() => {
     const storagedCart = localStorage.getItem('@RocketShoes:cart');
 
@@ -35,25 +36,30 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
   
   useEffect(() => {
     (async () => {
-      const { data } = await api.get('stock');
-      setStock(data);
+      const { data: dataStocks } = await api.get('stock');
+      const { data: dataProducts } = await api.get('products');
+      setStocks(dataStocks);
+      setProducts(dataProducts);
     })();
   }, []);
 
   const addProduct = async (productId: number) => {
     try {
-      const stockProduct = stock.find((stock: Stock) => stock.id === productId);
+      const isProductExist = products?.find((product) => product.id === productId);
+      if (!isProductExist) {
+        toast.error('Erro na adição do produto');
+        return;
+      }
+      const stockProduct = stocks.find((stock: Stock) => stock.id === productId);
       let dataCart = cart;
-      const isExists = cart?.find((product) => product.id === productId);
-      if (isExists) {
-        if (!isExists || !stockProduct || isExists?.amount + 1 > stockProduct?.amount) {
+      const isCartExist = cart?.find((product) => product.id === productId);
+      if (isCartExist) {
+        if (!isCartExist || !stockProduct || isCartExist?.amount + 1 > stockProduct?.amount) {
           toast.error('Quantidade solicitada fora de estoque');
           return;
         }
-        stock.find((stock: Stock) => stock.id === productId);
-        console.log(stock);
         dataCart = dataCart.map((product) => {
-          if (product.id === isExists?.id) {
+          if (product.id === isCartExist?.id) {
             return {
               ...product,
               amount: product.amount += 1,
@@ -61,6 +67,7 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
           }
           return product;
         });
+        localStorage.setItem('@RocketShoes:cart', JSON.stringify(dataCart));
       } else {
         const { data } = await api.get('products');
         const productData = data.find((product: Product) => product.id === productId);
@@ -69,9 +76,9 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
           ...productData,
           amount: 1
         }]
+        localStorage.setItem('@RocketShoes:cart', JSON.stringify(dataCart));
       }
       setCart(dataCart);
-      localStorage.setItem('@RocketShoes:cart', JSON.stringify(dataCart));
     } catch {
       toast.error('Erro na adição do produto');
     }
@@ -79,9 +86,14 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
   
   const removeProduct = (productId: number) => {
     try {
+      const isCartExist = cart?.find((product) => product.id === productId);
+      if (!isCartExist) {
+        toast.error('Erro na remoção do produto');
+        return;
+      }
       let dataCart = cart.filter((product) => product?.id !== productId);
-      setCart(dataCart);
       localStorage.setItem('@RocketShoes:cart', JSON.stringify(dataCart));
+      setCart(dataCart);
     } catch {
       toast.error('Erro na remoção do produto');
     }
@@ -92,15 +104,21 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
     amount,
   }: UpdateProductAmount) => {
     try {
-      const stockProduct = stock.find((stock: Stock) => stock.id === productId);
+      if (amount < 1) return;
+      const isProductExist = products?.find((product) => product.id === productId);
+      if (!isProductExist) {
+        toast.error('Erro na alteração de quantidade do produto');
+        return;
+      }
+      const stockProduct = stocks.find((stock: Stock) => stock.id === productId);
       let dataCart = cart;
-      const isExists = cart?.find((product) => product.id === productId);
-      if (!isExists || !stockProduct || isExists?.amount > stockProduct?.amount) {
+      const isCartExist = cart?.find((product) => product.id === productId);
+      if (!isCartExist || !stockProduct || isCartExist?.amount > stockProduct?.amount) {
         toast.error('Quantidade solicitada fora de estoque');
         return;
       }
       dataCart = dataCart.map((product) => {
-        if (product.id === isExists?.id) {
+        if (product.id === isCartExist?.id) {
           return {
             ...product,
             amount: amount
@@ -108,8 +126,8 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
         }
         return product;
       });
-      setCart(dataCart);
       localStorage.setItem('@RocketShoes:cart', JSON.stringify(dataCart));
+      setCart(dataCart);
     } catch {
       toast.error('Erro na alteração de quantidade do produto');
     }
